+++ Sources/Entities/PlayerStateInfo.swift
@@ -10,7 +10,7 @@ import Foundation
 
 public struct PlayerStateInfo: Codable {
     /// Gets or sets the index of the now playing audio stream.
-    public var audioStreamIndex: Int?
+    public var audioStreamIndex: Int32?
     /// Gets or sets a value indicating whether this instance can seek.
     public var canSeek: Bool?
     /// Gets or sets a value indicating whether this instance is muted.
@@ -24,27 +24,15 @@ public struct PlayerStateInfo: Codable {
     /// Gets or sets the play method.
     public var playMethod: PlayMethod?
     /// Gets or sets the now playing position ticks.
-    public var positionTicks: Int?
+    public var positionTicks: Int64?
     /// Gets or sets the repeat mode.
     public var repeatMode: RepeatMode?
     /// Gets or sets the index of the now playing subtitle stream.
-    public var subtitleStreamIndex: Int?
+    public var subtitleStreamIndex: Int32?
     /// Gets or sets the volume level.
-    public var volumeLevel: Int?
+    public var volumeLevel: Int32?
 
-    public init(
-        audioStreamIndex: Int? = nil,
-        canSeek: Bool? = nil,
-        isMuted: Bool? = nil,
-        isPaused: Bool? = nil,
-        liveStreamID: String? = nil,
-        mediaSourceID: String? = nil,
-        playMethod: PlayMethod? = nil,
-        positionTicks: Int? = nil,
-        repeatMode: RepeatMode? = nil,
-        subtitleStreamIndex: Int? = nil,
-        volumeLevel: Int? = nil
-    ) {
+    public init(audioStreamIndex: Int32? = nil, canSeek: Bool? = nil, isMuted: Bool? = nil, isPaused: Bool? = nil, liveStreamID: String? = nil, mediaSourceID: String? = nil, playMethod: PlayMethod? = nil, positionTicks: Int64? = nil, repeatMode: RepeatMode? = nil, subtitleStreamIndex: Int32? = nil, volumeLevel: Int32? = nil) {
         self.audioStreamIndex = audioStreamIndex
         self.canSeek = canSeek
         self.isMuted = isMuted
