+++ Sources/Entities/RemoteSubtitleInfo.swift
@@ -11,9 +11,9 @@ import Foundation
 public struct RemoteSubtitleInfo: Codable, Identifiable {
     public var author: String?
     public var comment: String?
-    public var communityRating: Double?
+    public var communityRating: Float?
     public var dateCreated: Date?
-    public var downloadCount: Int?
+    public var downloadCount: Int32?
     public var format: String?
     public var id: String?
     public var isHashMatch: Bool?
@@ -21,19 +21,7 @@ public struct RemoteSubtitleInfo: Codable, Identifiable {
     public var providerName: String?
     public var threeLetterISOLanguageName: String?
 
-    public init(
-        author: String? = nil,
-        comment: String? = nil,
-        communityRating: Double? = nil,
-        dateCreated: Date? = nil,
-        downloadCount: Int? = nil,
-        format: String? = nil,
-        id: String? = nil,
-        isHashMatch: Bool? = nil,
-        name: String? = nil,
-        providerName: String? = nil,
-        threeLetterISOLanguageName: String? = nil
-    ) {
+    public init(author: String? = nil, comment: String? = nil, communityRating: Float? = nil, dateCreated: Date? = nil, downloadCount: Int32? = nil, format: String? = nil, id: String? = nil, isHashMatch: Bool? = nil, name: String? = nil, providerName: String? = nil, threeLetterISOLanguageName: String? = nil) {
         self.author = author
         self.comment = comment
         self.communityRating = communityRating
