+++ Sources/Entities/TaskTriggerInfo.swift
@@ -13,21 +13,15 @@ public struct TaskTriggerInfo: Codable {
     /// Gets or sets the day of week.
     public var dayOfWeek: DayOfWeek?
     /// Gets or sets the interval.
-    public var intervalTicks: Int?
+    public var intervalTicks: Int64?
     /// Gets or sets the maximum runtime ticks.
-    public var maxRuntimeTicks: Int?
+    public var maxRuntimeTicks: Int64?
     /// Gets or sets the time of day.
-    public var timeOfDayTicks: Int?
+    public var timeOfDayTicks: Int64?
     /// Gets or sets the type.
     public var type: String?
 
-    public init(
-        dayOfWeek: DayOfWeek? = nil,
-        intervalTicks: Int? = nil,
-        maxRuntimeTicks: Int? = nil,
-        timeOfDayTicks: Int? = nil,
-        type: String? = nil
-    ) {
+    public init(dayOfWeek: DayOfWeek? = nil, intervalTicks: Int64? = nil, maxRuntimeTicks: Int64? = nil, timeOfDayTicks: Int64? = nil, type: String? = nil) {
         self.dayOfWeek = dayOfWeek
         self.intervalTicks = intervalTicks
         self.maxRuntimeTicks = maxRuntimeTicks
